pyproject.toml
[project]
name = "meta"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "matplotlib>=3.10.7",
    "numpy<2",
    "pandas>=2.3.3",
    "pyyaml>=6.0.3",
    "rdkit>=2025.9.1",
    "requests>=2.32.5",
    "scikit-learn>=1.7.2",
    "seaborn>=0.13.2",
    "torch==2.4.1+cu121",
    "torch-geometric>=2.7.0",
    "torchaudio==2.4.1+cu121",
    "torchvision==0.19.1+cu121",
]

[tool.uv.sources]
torch = { index = "pytorch-cu121" }
torchvision = { index = "pytorch-cu121" }
torchaudio = { index = "pytorch-cu121" }

[[tool.uv.index]]
name = "pytorch-cu121"
url = "https://download.pytorch.org/whl/cu121"
explicit = true
